CDR-934 test for AqlQueryParser::parsePath, move utils to PredicateHelper
